hexes are now again unsigned long, francinette pass

+++ put_functions/ft_puthex_caps.c
@@ -6,13 +6,13 @@
 /*   By: amaula <amaula@student.hive.fi>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/12 15:57:17 by amaula            #+#    #+#             */
-/*   Updated: 2024/05/13 12:35:53 by amaula           ###   ########.fr       */
+/*   Updated: 2024/05/13 12:57:50 by amaula           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../ft_printf.h"
 
-int	ft_puthex_caps(unsigned int hc)
+int	ft_puthex_caps(unsigned long hc)
 {
 	char	digit;
 	int 	ret;
